+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2023-04-27
+
+### Added
+
+- `placeholder` property to set `placeholder` in `<Input />` components
+
 ### Security
 
 - Dependabot: Bump vm2 from 3.9.16 to 3.9.17
@@ -125,10 +131,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/0.9.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/0.10.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
+[0.10.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.8.1...0.9.0
 [0.8.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.8.0...0.8.1
 [0.8.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.7.0...0.8.0